Code cleanup

+++ src/helper/RandString.h
@@ -16,7 +16,7 @@ inline std::string randomString() {
         std::random_device dev;
         std::mt19937 rng(dev());
         std::uniform_int_distribution<std::mt19937::result_type> dist(1, INT32_MAX);
-        const int roll = dist(rng);
+        const unsigned long roll = dist(rng);
         tmp_s += alphanum[roll % (sizeof(alphanum) - 1)];
     }
